requirements/base.txt
tensorflow==2.6.0
Django==3.2.6
django-environ==0.4.5
djangorestframework==3.12.4
django_compressor==2.4.1
django-htmlmin==0.11.0
pymorphy2==0.9.1
dill>=0.3.4
pydantic>=1.8.2
cairosvg==2.5.2
PyYAML~=5.4.1
transliterate==1.10.2
imgaug==0.4.0
dict-recursive-update==1.0.1
pydub==0.25.1
psutil==5.8.0
pynvml==11.0.0
moviepy==1.0.3
numpy==1.19.5
albumentations==1.0.3
opencv-python==4.5.3.56
python-dotenv==0.19.0
